requirements.txt
# GUI Framework
PyQt5>=5.15.0

# System Tray Integration
pystray>=0.19.0

# Clipboard Monitoring
pyperclip>=1.8.0

# Global Hotkey Support
keyboard>=0.13.0; sys_platform == 'win32'
pynput>=1.7.0; sys_platform == 'linux'

# Image Processing
Pillow>=9.0.0

# Windows-specific clipboard access
pywin32>=300; sys_platform == 'win32'

# Database
# SQLite is included in Python standard library

# Additional utilities
python-dateutil>=2.8.0
